requirements.txt
datasets
peft
trl
tiktoken
accelerate
einops
bitsandbytes
huggingface_hub
python-dotenv